fix: Prevents projectIds from being reset when an insert takes place
Cause:

When plugin is loaded, it would load the project id.

conductor-cli would then increment it externally.

When the insert task command is called, it doesn't reload the settings
but uses what is loaded when the plugin starts.

Then when it saves at the end, it saves it with the old values,
resetting the changes performed externally.

Solution:

Settings are now loaded at the start of the command.